leader-worker: optimize code and fix multiple register of internal listeners
